moved footer file
