+++ eCommerce-app/src/layouts/Layout.vue
@@ -9,7 +9,7 @@
 </template>
 <script setup lang="ts">
 import Header from "../shared/components/molecules/Header.vue";
-import Footer from "../components/navigation/Footer.vue";
+import Footer from "../shared/components/molecules/Footer.vue";
 </script>
 <style>
 html,
